requirements.txt
asgiref==3.3.4
authentication==1.1.0
click==8.0.1
colorama==0.4.4
Django==3.2.4
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
passlib==1.7.4
Pillow==8.2.0
pycryptodome==3.10.1
PyJWT==2.1.0
pytz==2021.1
redis==3.5.3
sqlparse==0.4.1
Werkzeug==2.0.1
WTForms==2.3.3
